% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_envelope.R
\name{adjust_envelope}
\alias{adjust_envelope}
\title{adjust_envelope}
\usage{
adjust_envelope(t, y, z, w, QC_flag, type = "y", trs = 0.5,
  alpha = 0.02, wmin = 0.2)
}
\arguments{
\item{t}{Date object, corresponding date of y. If specified, yearly $ylu$ will
be calculated, which is used to filter peak season.}

\item{y}{Numeric vector, vegetation index time-series}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{QC_flag}{Quality Control (QC) flag of original vegetation time-series y.}

\item{type}{Character, ylu get by "y" or "z".}

\item{alpha}{Double value in \code{[0,1]}, quantile prob of ylu_min.}

\item{wmin}{Double, minimum weight of bad points, which could be smaller
the weight of snow, ice and cloud.}

\item{ylu}{Numeric, lower and upper boundary of y.}
}
\description{
Approach to upper envelope. This function works site by site.
}
\details{
Smoothed time-series by Whittaker is often underestimated in the peak of
growing season (abbreviated as peak season). In R, we can further constrain
by yearly ylu. But, it's difficult to implement in GEE.
\enumerate{
\item Not outlier.
\item In middle growing season.
\item only good values.
}
}
