% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Whittaker.R
\name{check_Whittaker}
\alias{check_Whittaker}
\title{check the performance of Whittaer, and compared with other methods}
\usage{
check_Whittaker(sites, df_trim, file, methods)
}
\description{
check the performance of Whittaer, and compared with other methods
}
