% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_fitting.R
\name{coef_roughness}
\alias{coef_roughness}
\title{Roughness of matrix}
\usage{
coef_roughness(x)
}
\arguments{
\item{x}{matrix, \verb{[ngrid, ntime]}}
}
\value{
\code{Roughness = diff(x)^2}
}
\description{
Roughness of matrix
}
\examples{
data("lst_noise")
coef_roughness(lst_noise$`50\%`$y)
}
