% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{d_ref}
\alias{d_ref}
\title{data.frame of Reference curve used to simulate noises.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 46 rows and 4 columns.
}
\usage{
d_ref
}
\description{
data.frame of Reference curve used to simulate noises.
}
\examples{
data("d_ref")
print(head(d_ref))
}
\seealso{
\link{simu_noise}
}
\keyword{datasets}
