% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSimuNoise_TSF.R
\name{fitSimuNoise_TSF}
\alias{fitSimuNoise_TSF}
\title{fitSimuNoise_TSF}
\usage{
fitSimuNoise_TSF(
  lst_noise_random,
  lst_noise_keypoint,
  nptperyear,
  nyear,
  indir,
  ...
)
}
\arguments{
\item{...}{ignored}
}
\description{
fitSimuNoise_TSF
}
