% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSimuNoise_WHIT.R
\name{fitSimuNoise_WHIT}
\alias{fitSimuNoise_WHIT}
\title{fitSimuNoise_WHIT}
\usage{
fitSimuNoise_WHIT(lst_noise, lambdas = NULL, nptperyear = 23,
  wmin = 0.2, ..., outdir = "./OUTPUT", is_save = FALSE,
  limits = NULL)
}
\arguments{
\item{lst_noise}{list of simulated random and keypoint noise, \code{list(list(y, QC))}}

\item{lambdas}{lamdas for Whittaker}

\item{nptperyear}{Integer, number of images per year.}

\item{wmin}{Double, minimum weight of bad points, which could be smaller
the weight of snow, ice and cloud.}

\item{...}{other parameters to \code{\link[=check_input]{check_input()}}, e.g. \code{perc_wc}, \code{wmin}.
\code{wmin} also is passed to \code{wFUN}.}

\item{limits}{If not null, only process the first \code{limits} pixel severing
to debug.}
}
\description{
fitSimuNoise_WHIT
}
\examples{
data('lst_noise')
r <- fitSimuNoise_WHIT(lst_noise)
}
\seealso{
\link{check_input}
}
