% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_fitting.R
\name{gof_fitting}
\alias{gof_fitting}
\title{Good-of-fitting (GOF) of fitted time-series}
\usage{
gof_fitting(mat_ref, mat_fit, I_rem, times = 100)
}
\arguments{
\item{mat_ref}{matrix of reference curvr \code{[ngrid, ntime]}}

\item{mat_fit}{matrix of fitted time-series \code{[ngrid, ntime]}}

\item{I_rem}{index of succeed fitted pixels.}

\item{times}{For each, how many times noise simulated?}
}
\description{
Good-of-fitting (GOF) of fitted time-series
}
\examples{
data("mat_ref")
data("lst_noise")
d_gof <- gof_fitting(mat_ref, lst_noise$real$y)
}
\seealso{
\link[phenofit:GOF]{phenofit::GOF}
}
