% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_optim_FUN.R
\name{lambda_optim}
\alias{lambda_optim}
\title{According to v-curve theory, get the optimal lambda value.}
\usage{
lambda_optim(
  sitename,
  df,
  deltaT,
  extend = T,
  IsPlot = F,
  IsSave = F,
  file = "test_whit_lambda.pdf",
  wFUN = wBisquare,
  iters = 2,
  is_normalize = FALSE
)
}
\arguments{
\item{deltaT}{int, nyears chunk}
}
\description{
Whittaker balanced the fidelity and smooth. The agreement index maybe poor
than others. But it'is much smoothing.
}
