% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{lst_noise}
\alias{lst_noise}
\title{Simulated random noise and keypoint noise.}
\format{
An object of class \code{list} of length 6.
}
\usage{
lst_noise
}
\description{
10\\%, 30\\% and 50\\% simulated random noise and 3 type keypoint noise.
}
\details{
Each noise contains two variables:
\itemize{
\item \strong{\code{y}:} vegetation enhanced index (EVI) time-series
\item \strong{\code{QC}:} in the range of \verb{[0-3]}, see \code{qc_summary}
}
}
\examples{
data("lst_noise")
str(lst_noise, 2)
}
\keyword{datasets}
