% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{mat_ref}
\alias{mat_ref}
\title{matirx of Reference curve used to evaluate GOF.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 23 columns.
}
\usage{
mat_ref
}
\description{
matirx of Reference curve used to evaluate GOF.
}
\examples{
data("mat_ref")
print(head(mat_ref))
}
\seealso{
\link{gof_fitting}
}
\keyword{datasets}
