% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_rough.R
\name{rough_fitting}
\alias{rough_fitting}
\title{rough_fitting}
\usage{
rough_fitting(
  sitename,
  df,
  st,
  .FUN = wWHIT,
  lambda = NULL,
  IsPlot = FALSE,
  print = FALSE,
  ...
)
}
\arguments{
\item{lambda}{Unless lambda is constant, lambda should be null.}
}
\description{
rough_fitting
}
