% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_noise.R
\name{simu_noise}
\alias{simu_noise}
\title{simu_noise}
\usage{
simu_noise(d, perc = 0.1, type = "random", times = 100)
}
\arguments{
\item{d}{data.frame, at least with the column of \code{ref} and \code{SummaryQA}.}

\item{type}{One of c("random", "real", "maxK", "maxDer")}

\item{times}{For each, how many times noise simulated?}
}
\description{
simulate specific type noise 100 times through \code{add_noise}.
}
\note{
Each time only process one site
}
\examples{
data("d_ref")
r <- simu_noise(d_ref, type = "real")
}
\seealso{
\link{add_noise}
}
