% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_WHIT.R
\name{smooth_WHIT}
\alias{smooth_WHIT}
\alias{smooth_WHIT_df}
\title{Whittaker smoother}
\usage{
smooth_WHIT(y, w, lambda, iters = 3, IsPlot = FALSE, ...)

smooth_WHIT_df(d, lambda, iters = 3, IsPlot = FALSE, ...)
}
\arguments{
\item{y}{Numeric vector, vegetation index time-series}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{...}{other parameters to \code{\link[=check_input]{check_input()}}, e.g. \code{perc_wc}, \code{wmin}.
\code{wmin} also is passed to \code{wFUN}.}
}
\description{
Whittaker smoother
}
\seealso{
\code{\link[=check_input]{check_input()}}
}
