% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittaker-package.R
\docType{package}
\name{whittaker-package}
\alias{whittaker}
\alias{whittaker-package}
\title{whittaker: Whittaker smoother}
\description{
Whittaker smoother.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
